fix: a11y, contrast, dead-token cleanup, DRY polish
- A11y: aria-current='location' (was string 'true') + aria-label on preview live region
- Contrast: --grade brighter green (#34D399) in dark mode for legibility on --surface
- Dead tokens: remove secondary/decorative/hovered from tailwind.config + --color-* backward-compat vars from globals.css (zero references confirmed)
- Visual: suppress bottom-left corner tick in FigurePlate when caption/fig is present
- DRY: extract SectionLabel to src/components/ui/section-label.tsx (education + awards)
- DRY: extract formatIsoDate to src/lib/format.ts (article-card + article slug page)
- DRY: gate article meta-chip wrapper so it renders only when metadata.lang is set
- DRY: remove redundant !d || guard in assertProjectMetadata second check
