+++ src/app/globals.css
@@ -44,13 +44,6 @@
   --grade:         #1E8A4C;
   --hand:          #C8402F; /* same as accent in light */
   --grid-line:     rgba(30,95,204,.07);
-  /* backward-compat aliases (used by not-yet-rewritten components) */
-  --color-primary:    var(--primary);
-  --color-accent:     var(--accent);
-  --color-secondary:  var(--muted);
-  --color-decorative: var(--muted);
-  --color-surface:    var(--surface);
-  --color-hovered:    var(--surface);
 }
 
 [data-mode='dark'] {
@@ -63,16 +56,9 @@
   --muted:         #8FA6B5;
   --primary:       #5BA8FF;
   --accent:        #FF6B5A;
-  --grade:         #1E8A4C; /* kept same; no dark override in brief */
+  --grade:         #34D399; /* brighter green for contrast on dark surface */
   --hand:          #FF9A6B;
   --grid-line:     rgba(56,189,248,.07);
-  /* backward-compat aliases */
-  --color-primary:    var(--primary);
-  --color-accent:     var(--accent);
-  --color-secondary:  var(--muted);
-  --color-decorative: var(--muted);
-  --color-surface:    var(--surface);
-  --color-hovered:    var(--surface);
 }
 
 body {
